[UPD] Add new macros for the BTK release number (BTK_VERSION_MAJOR, BTK_VERSION_MINOR, BTK_VERSION_PATCH). [UPD] Remove some CL warnings (size_t vs int). [FIX] Wrong description in btkPWRFileIO.cpp to explain the content of a PWR file.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@353 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/BasicFilters/btkForcePlatformsExtractor.cpp
@@ -317,7 +317,7 @@ namespace btk
     {
       pValue = pCalMatrix->GetInfo();
       ForcePlatform::CalMatrix cal = fp->GetCalMatrix();
-      if (static_cast<size_t>(pValue->GetValues().size()) >= (coefficientsAlreadyExtracted + cal.size()))
+      if (pValue->GetValues().size() >= (coefficientsAlreadyExtracted + cal.size()))
       {
         for (int i = 0 ; i < cal.cols() ; ++i)
           for (int j = 0 ; j < cal.rows() ; ++j)
